third_party/WebKit/LayoutTests/external/wpt/mixed-content/allowed/http-csp/same-host-https/audio-tag/top-level/keep-scheme-redirect/allowed.https-expected.txt

This is a testharness.js-based test.
FAIL opt_in_method: http-csp
                                 origin: same-host-https
                                 source_scheme: https
                                 context_nesting: top-level
                                 redirection: keep-scheme-redirect
                                 subresource: audio-tag
                                 expectation: allowed assert_equals: The triggered event should match 'allowed'. expected "allowed" but got "blocked"
Harness: the test ran to completion.
